Improve options toggler for playlist/albums items
Wrap the glyphicon inside a `.options-toggler` div to better catch
it through jQuery and stylize it.

Also we are more consistent with the video or music boxes' mark-up.

The icon is now an horizontal option rather than a vertical one since
the options will be displayed on the same line.