Unify the `title` elements' rules
Move the `.title` rule to the `.basic-box` level so more elements will
pick this rule and this avoid mixing-in rules in other elements and this
is DRY-er.

Also remove the `padding-right` rule on some of these elements as this
is no longer necessary since the `options-toggler` div is never hidden,
only its glyphicon so it always has a width.