Pass the playlist/album when clicking options
Sometimes, to find a particular record, we need to tell the model in
which playlist or album it needs to look to find it.

Thus, when we click on an option on an item inside a playlist, we need to
pass the latter as well to the called function.

Also move the associated event handler in the right file since it is
related to medias, and not specifically to the container.