Improve the SoundCloud stream look
Improve the look of the SoundCloud stream by relying on a new style
of div for musics (i.e. `.music.bar`) to have a look close to the
actual SoundCloud interface and add a timeline with icons describing
the activity, again, like the SoundCloud interface.

Playlists are no longer displayed inside `.musid.box` elements so people
can listen to them, it was not possible before. They are exactly like
any other playlist element inside the application.

Also, the `.music.basic-box` was a first try to match the SoundCloud
interface but with the addition of the `.music.bar` element, it is no
longer necessary.

The activities are special cased when they are transformed to records
inside the `fetch` method of the SoundCloud model since the record
itself is contained either inside the `track` or the `playlist` field
but we also need access to the `user` and `type` fields of the root JSON
hash to display it to the user.