Make sure that we are only dealing with Record objects
Since we want to compute the `human_time` field in-place and no longer
store it in memory, we need to make sure that we are only dealing with
instances of Record and not just raw JSON hash.

Thus, we need to move the record conversion from the local model's
`readFiles` methods to the `files` one since `process.send` convert
things to JSON and it doesn't round-trip dealing with custom objects
so we get raw hashes instead of record objects.

Also, we need to convert playlists' entries to record objects We need to
introduce a `Record#raw` static method, just to copy fields in-place
because the database stores records that already have the field names
that we use as the attributes record objects since we may be dealing
with data coming from SoundCloud, YouTube or the local file system.