Avoid creating the options div for video boxes
Historically, we used to generate the HTML for `.options` div for every
single box. Then, this was changed to generate the HTML on demand to
speed up a bit the render process since people are not going to download
or share every single element.

To avoid generating twice the HTML, we check whether the current box has
an `.option` element and only generate if it's not the case. However, I
forgot to remove the legacy code for YouTube video boxes so the options
were empty.