Unify the interface to pick a file/folder
Create a `FilePicker` class with an `open` method to easily add file
selector inside the application. The allowed kind of dialogs are
'picture' or 'folder'.

Thus, we can also fix the configuration page to correctly display the
dialog as it was not the case before since `dialog` was not defined.
Also rely on this new API for tagging dialogs so people can choose the
cover file of their musics.