Set-up the basis of the dashboard section
Allow the `meta/dashboard` to be routed to different actions so we can
have a dashboard with a navbar and different sections and clicking on
the dashboard icon will redirect to the last action chosen by the user.

This follows the pattern of `local/files` for instance, where the user
can either choose the singles, albums or artists view.