Share the playlist partials
The SoundCloud and local playlists markup is almost identical; on the
other hand, YouTube has a pretty different style since it is displayed
on a single page but it still close.

So let's move all the markup inside a single partial and add a bit of
logic to the `Record` object and the view to correctly generate the
appropriate HTML code.

This avoid to have an helper and a view for each service plus there are
less cache entries for templates in memory.