Improve the full-screen mode for videos
Now, we apply the full-screen call on the `.player_frame` element
rather than on the `.player_frame_container` one to avoid having
the white border when in full-screen mode.

The `.controls` element has been moved inside the former to make sure
that they are present when full-screen is enabled. Also, the controls
are now automatically hidden when the mouse hasn't moved for 2 seconds.

The controls in general have been improved; the mouse is hidden as well
when the controls are and the icon to switch the mode automatically
change regarding the current displaying.

The API to manage the full-screen mode is now living inside the Ui
object to make the code clearer and make us able to reuse it elsewhere.

Finally, we now ensure that the interface is correctly set-up when the
user exits it using "Esc" or "F11".