Compile the CSS file manually
Avoid having Less loaded inside our page to reduce the boot time of
the application.

Now, when a modification is done on the CSS, we just need to run
`make css` and reload the page.

Moreover, we can remove the `assets/javascripts/less.js` file which was
pretty naughty when we were using grep inside the folder.