Improve the checkboxes for searching
Rely on custom elements (i.e. switches) for picking the services the
user want to search on, this looks a bit nicer and it takes less place
on the screen than the previous mark-up.

The algorithm to know which services have been picked remain almost the
same, we are just relying on an hidden input rather than a checkbox.