Widely improve the search feature
Allow users to specify whether they want to search for a specific media
inside their collection or rather on the chosen service(s). This makes
searches way faster if people don't want results coming from the
internet.

Also unify the search interface of all the different services. Move all
the `search` method to the associated service object and correctly
dispatch to the model then on.

The YouTube model doesn't have a search on a per-field basis for now as
they aren't enough things set-up and because the artist is part of the
title most of the times.

Also update some useless code; in the local model, let's unify the
`findBy` methods under a single one just like the SoundCloud model.
On the SoundCloud side, let's update the net search to use the new
version of the API.

This commit is just the first part of a larger refactoring.