Allow the click on the progression
Since we can't stylize through CSS the left hand side of a range input,
we had to add a `.progression` div that would be displayed atop the
input showing the current progression.

The problem was that we had no event handler clicking on this element so
even though the hit box of the range input is pretty large, if the user
precisely click on the two pixels of this bar, the current time wouldn't
change.