Correctly display the SoundCloud user's playlists
We were previously only fetching the five first items of the user's
playlists.

If we don't want to fetch items for each playlist, we need to fallback
to the V1 SoundCloud API as we want to distinguish between owned and
liked playlists but V2 has a shared endpoint for both of them.