Add links to artist's page on local media boxes
Let's add a link on the local music boxes to go to the music's artist's
page so that people don't have to go to the local/artists action and
click on the artist to see their other medias.

Thus, we need to change the previous `a.artist` selector to a new
`.artist-thumbnail` element so that there is no conflict for styling.

Also, let's correctly position the options toggler inside media boxes
rather than relying on a tricky "&nbsp;" hack to make sure that it gets
at the bottom of the box.