Implement the YouTube history ourself
Since the YouTube Data API version 3 doesn't give us access to the
user's history anymore, let's implement this feature ourself.

The online history and the one shown here won't match anymore but this
is far easier for us, plus the history endpoint was very slow anyway
and thus, people may have a "music" history showing only videos that
have been watched with this software and still have an history with
their podcasts, series, you name it.

Under the hood, the history is stored like any local playlist, in
JSPF, on the disk.

We are also now adding played elements to the history; this was not the
case before.