Display progression reading local music files
This operation may take a lot of time so let's display the current
progress specifying the number of files that have already been processed
and the total of files to process.

The Ui#hideLoader method now also hides the `.loader-text` element to
make sure that the message disappears when showing the loader for
another action.