Remove the pop-up to create local playlists
The creation interface will be unified later for all the services
under the dialog that allows the user to add elements to playlists.

This pop-up element wasn't really useful to other parts of the
application, thus this wasn't consistent with the rest of the user
interface and having a navigation bar with only a plus sign feels a
bit ridiculous.