Allow creation of playlists
Add an input inside the dialog that allows the user to add elements to
playlists to create brand new playlist.

If the user picks a different tab inside the dialog, the playlist will
be created on the corresponding service.

We need to have a consistent API for the `createPlaylist` methods; the
playlist record should be returned by the promise to be able to create
the option inside the dialog appropriately. Thus the local model also
received some changes.