Hide the main shadow pressing esc inside an input
The user should be able to hide the main shadow pressing the escape key
even though it is inside a input.

This feature was was added in 036bdad8 and the check has always been
there but this was to avoid any conflict pressing escape inside the
search field but this can't happen as the search bar isn't clickable
when the main shadow is visible.