Make medias hold a reference to their set
We used to return a hash containing a `record` field and a `playlist`
field dealing with records that were stored inside playlists or albums.

The problem is that this allocates a new object each time we want to
find a record and this is terribly inefficient since we already have
access to the set instantiating its medias or the other way around.

This also avoids some weird checks to see if we are dealing with a media
from a set or just from a raw collection.

Also now, to be consistent across the different services, YouTube
playlists are also holding their items rather than storing them inside a
specific hash inside the cache. But since fetching a playlist and its
items is done through two different requests, we still need to keep
track of the playlists that have their items already loaded.