Make the progression range scrollable to zero
The progression range couldn't be dragged to the very beginning if the
mouse was released outside of the `.duration` element since the `click`
event handler of the latter wasn't triggered.

Also the `Player#goTo` method resets the value of the `auto_progression`
attribute since this method is called whenever the timing is sort of
"confirmed".