Make downloading API more consistent
Rather than dealing with different APIs for downloading medias, let's
rather just pass a hash that contains the needed attributes to the
different methods.

With this change, we can now easily collect the different values filled
in the form through the jQuery's `extractFields` method introduced in
this commit.

Furthermore, a lot of our APIs are already dealing with hashes rather
than different parameters so this ease inner calls and avoid extra
allocations.