Look in the right place trying to play singles
Trying to play a single on an artist's page or inside search results
would lead to an error as we are roughly loading the current action's
cache and trying to look for a record.

This can't be that simple though because `search_results` and `artist`
are "meta" actions that are just gathering some data of other actions.

This is not a problem with sets and albums as we are passing the set
to the finder method so we know exactly where to look.