Cache local music tags on the disk
Processing a huge folder can take several seconds so let's rather
save the tags that are already known on the disk.

The list of files is saved as well as the library so that we can infer
which files have been removed and which files have been added.

The `TaggingService#extract` method now relies on the fact that the
`get` method from the tagging library push files inside an existing
library to process new files.

It delegates to the `TaggingService#remove` method to remove files
that are no longer present on the disk.