Implement the repeat mode in terms of native features
The video and audio elements have a `loop` attribute that allows
repeating them when they're finished. Let's rather rely on this rather
than on an ugly check in our code.

This also fixes a problem when the loop mode was turned on for a music
but was kept on clicking on another one, even though the user interface
was properly reset.