Improve the playing section of the player bar
Gather the style of the current played media's details under a new
element with the class `.media-details` to share the CSS rules easily
between the elements of the current set and the current media.

Also improve a bit the styling of the current set and give better names
inside the mark-up as it was really messy/unclear.