Some improvements to the tagging service
Make the code up-to-date as several architectural changes have been
brought to the tagging library.

There's now a built-in `Library` object that implements `get` to easily
get a brand new library or update an existing one and that also has
a `remove` method to remove references to some files.

Also cleaned up a bit the code.