Totally change the way we find and play medias
Instead of searching through the whole user's collection, let's rather
keep a in-memory copy of the last rendered view's context as we know
that the user can only play medias from it.

This widely speeds up finding records.

Therefore, since we have a copy of the view context, we can determine
the playing queue very easily by "flattening" the context. This also
avoids the ugly doubly linked list that we used to have to maintain.

Thus, all the finders method can be removed from the different models
as collection are now smart enough to know how to find elements or
generate a flattened version of them. If we are passing raw objects to
the view, they are automatically converted to `Context` instances which
know how to generate the collection of medias.