Load all the SoundCloud playlists under the same action
The SoundCloud API gives access to the liked and owned playlists under
the same endpoint so it's a bit hard to maintain two different actions
to make the distinction between them.

Also, we needed to rely on the version 1 of the API to only load owned
playlists, which wasn't very clean.