Handle ISO 8601 time conversion in its own function
Rather than polluting the `Timing#time` function, let's rather move
the conversion from ISO 8601 string to its own `fromISO8601` function.

Since the former is really frequently called, we have a tiny performance
boost since we avoid an extra check each time we call it. Also, since
we are 90% of time calling it with numbers and since we know where
we are dealing with ISO strings, it's fairly easy to use this new
method.