Remove the equalizer code
At least for now since it's badly implemented and produces a memory
leak as the sampled array seems to never to garbage collected as
musics are read (sometimes the memory used was ~800 megabytes after
a few hours of use).

For now, let's rather display a "play" icon to still easily see where
the current media is played from.