Avoid using links for the sign-in / sign-out
Links should only be used for the rendering logic ; here, we are
rather displaying a window and listening to various events. There is
no rendering involved.

Also introduce an `Actions` class that handles the different actions
that are present in the application. Let's use a `data-action` syntax
to automatically bind the click on certain elements to an action.