Refactor the controller stack
Now, controllers are instantiated once and we access their methods
to render an action rather than calling a static method.

Also, since there is a clear pattern that emerge from controllers,
let's simplify the way we define actions to make the controller
files as clear as possible.

Finally, remove the `Controller.for` mapping which is far from ideal.